+++ resources/views/bootstrap-4/form-radio.blade.php
@@ -4,10 +4,9 @@
 
         @if($isWired())
             wire:model{!! $wireModifier() !!}="{{ $name }}"
-        @else
-            name="{{ $name }}"
         @endif
 
+        name="{{ $name }}"
         value="{{ $value }}"
 
         @if($checked)