+++ resources/views/tailwind-2/form-select.blade.php
@@ -5,17 +5,28 @@
         <select
             @if($isWired())
                 wire:model{!! $wireModifier() !!}="{{ $name }}"
-            @else
-                name="{{ $name }}"
             @endif
+
+            name="{{ $name }}"
 
             @if($multiple)
                 multiple
             @endif
 
+            @if($placeholder)
+                placeholder="{{ $placeholder }}"
+            @endif
+
             {!! $attributes->merge([
                 'class' => ($label ? 'mt-1' : '') . ' block w-full'
             ]) !!}>
+
+            @if($placeholder)
+                <option value="" disabled @if($nothingSelected()) selected="selected" @endif>
+                    {{ $placeholder }}
+                </option>
+            @endif
+
             @forelse($options as $key => $option)
                 <option value="{{ $key }}" @if($isSelected($key)) selected="selected" @endif>
                     {{ $option }}