+++ src/Components/FormSelect.php
@@ -16,6 +16,8 @@ class FormSelect extends Component
     public $options;
     public $selectedKey;
     public bool $multiple;
+    public bool $floating;
+    public string $placeholder;
 
     /**
      * Create a new component instance.
@@ -30,12 +32,15 @@ class FormSelect extends Component
         $default = null,
         bool $multiple = false,
         bool $showErrors = true,
-        bool $manyRelation = false
+        bool $manyRelation = false,
+        bool $floating = false,
+        string $placeholder = ''
     ) {
         $this->name         = $name;
         $this->label        = $label;
         $this->options      = $options;
         $this->manyRelation = $manyRelation;
+        $this->placeholder  = $placeholder;
 
         if ($this->isNotWired()) {
             $inputName = static::convertBracketsToDots(Str::before($name, '[]'));
@@ -51,6 +56,7 @@ class FormSelect extends Component
 
         $this->multiple   = $multiple;
         $this->showErrors = $showErrors;
+        $this->floating   = $floating && !$multiple;
     }
 
     public function isSelected($key): bool
@@ -60,5 +66,14 @@ class FormSelect extends Component
         }
 
         return in_array($key, Arr::wrap($this->selectedKey));
+    }
+
+    public function nothingSelected(): bool
+    {
+        if ($this->isWired()) {
+            return false;
+        }
+
+        return is_array($this->selectedKey) ? empty($this->selectedKey) : is_null($this->selectedKey);
     }
 }