+++ src/Components/HandlesBoundValues.php
@@ -2,9 +2,12 @@
 
 namespace ProtoneMedia\LaravelFormComponents\Components;
 
+use DateTimeInterface;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
+use Illuminate\Support\Carbon;
+use Illuminate\Support\Str;
 use ProtoneMedia\LaravelFormComponents\FormDataBinder;
 
 trait HandlesBoundValues
@@ -52,9 +55,60 @@ trait HandlesBoundValues
 
         $bind = $bind ?: $this->getBoundTarget();
 
-        return $this->manyRelation
-            ? $this->getAttachedKeysFromRelation($bind, $name)
-            : data_get($bind, $name);
+        if ($this->manyRelation) {
+            return $this->getAttachedKeysFromRelation($bind, $name);
+        }
+
+        $boundValue = data_get($bind, $name);
+
+        if ($bind instanceof Model && $boundValue instanceof DateTimeInterface) {
+            return $this->formatDateTime($bind, $name, $boundValue);
+        }
+
+        return $boundValue;
+    }
+
+    /**
+     * Formats a DateTimeInterface if the key is specified as a date or datetime in the model.
+     *
+     * @param \Illuminate\Database\Eloquent\Model $model
+     * @param string $key
+     * @param DateTimeInterface $date
+     * @return void
+     */
+    private function formatDateTime(Model $model, string $key, DateTimeInterface $date)
+    {
+        if (!config('form-components.use_eloquent_date_casting')) {
+            return $date;
+        }
+
+        $cast = $model->getCasts()[$key] ?? null;
+
+        if (!$cast || $cast === 'date' || $cast === 'datetime') {
+            return Carbon::instance($date)->toJSON();
+        }
+
+        if ($this->isCustomDateTimeCast($cast)) {
+            return $date->format(explode(':', $cast, 2)[1]);
+        }
+
+        return $date;
+    }
+
+    /**
+     * Determine if the cast type is a custom date time cast.
+     *
+     * @param  string  $cast
+     * @return bool
+     */
+    protected function isCustomDateTimeCast($cast)
+    {
+        return Str::startsWith($cast, [
+            'date:',
+            'datetime:',
+            'immutable_date:',
+            'immutable_datetime:',
+        ]);
     }
 
     /**