+++ tests/Feature/views/date-time-casts.blade.php
@@ -1,0 +1,13 @@
+<x-form>
+    @bind($model)
+        <x-form-input name="date_a" />
+        <x-form-input name="date_b" />
+        <x-form-input name="date_c" />
+        <x-form-input name="date_d" />
+        <x-form-input name="date_e" />
+        <x-form-input name="date_f" />
+        <x-form-input name="date_g" />
+
+        <x-form-submit />
+    @endbind
+</x-form>