+++ tests/Feature/views/select-placeholder.blade.php
@@ -1,0 +1,5 @@
+<x-form>
+    <x-form-select name="select" placeholder="Choose..." :options="['a' => 'a', 'b' => 'b']"></x-form-select>
+
+    <x-form-submit />
+</x-form>